api/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
tensorflow==2.11.0
numpy==1.22.4
Pillow==9.4.0
protobuf==3.20.0